+++ rocket-chat-android-widgets/src/main/java/chat/rocket/android/widget/helper/Linkify.java
@@ -1,16 +1,8 @@
 package chat.rocket.android.widget.helper;
 
-import android.content.ClipData;
-import android.content.ClipboardManager;
-import android.content.Context;
-import android.content.Intent;
-import android.net.Uri;
 import android.text.SpannableString;
 import android.text.Spanned;
-import android.text.style.ClickableSpan;
-import android.view.View;
 import android.widget.TextView;
-import android.widget.Toast;
 import org.nibor.autolink.LinkExtractor;
 import org.nibor.autolink.LinkSpan;
 import org.nibor.autolink.LinkType;
@@ -41,33 +33,10 @@ public class Linkify {
       final int idx2 = link.getEndIndex();
       final String url = text.subSequence(idx1, idx2).toString();
 
-      spannableString.setSpan(createLinkSpan(url), idx1, idx2, Spanned.SPAN_EXCLUSIVE_EXCLUSIVE);
+      spannableString.setSpan(MarkDown.createLinkSpan(url), idx1, idx2,
+          Spanned.SPAN_EXCLUSIVE_EXCLUSIVE);
     }
 
     return spannableString;
-  }
-
-  private static ClickableSpan createLinkSpan(final String url) {
-    return new ClickableSpan() {
-      @Override
-      public void onClick(View view) {
-        final Context context = view.getContext();
-        try {
-          Intent intent = new Intent(Intent.ACTION_VIEW, Uri.parse(url));
-          intent.setFlags(Intent.FLAG_ACTIVITY_NEW_TASK);
-          context.startActivity(intent);
-          return;
-        } catch (Exception exception) {
-        }
-
-        try {
-          ClipboardManager clipboardManager =
-              (ClipboardManager) context.getSystemService(Context.CLIPBOARD_SERVICE);
-          clipboardManager.setPrimaryClip(ClipData.newPlainText("linkURL", url));
-          Toast.makeText(context, "Copied to clipboard", Toast.LENGTH_SHORT).show();
-        } catch (Exception exception) {
-        }
-      }
-    };
   }
 }
