+++ rocket-chat-android-widgets/src/main/java/chat/rocket/android/widget/helper/MarkDown.java
@@ -1,12 +1,20 @@
 package chat.rocket.android.widget.helper;
 
+import android.content.ClipData;
+import android.content.ClipboardManager;
+import android.content.Context;
+import android.content.Intent;
 import android.graphics.Typeface;
+import android.net.Uri;
 import android.text.SpannableString;
 import android.text.Spanned;
 import android.text.style.AbsoluteSizeSpan;
+import android.text.style.ClickableSpan;
 import android.text.style.StrikethroughSpan;
 import android.text.style.StyleSpan;
+import android.view.View;
 import android.widget.TextView;
+import android.widget.Toast;
 
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
@@ -22,6 +30,8 @@ public class MarkDown {
   public static void apply(TextView textView) {
     SpannableString text = new SpannableString(textView.getText());
     removeImage(text);
+    highlightLink1(text);
+    highlightLink2(text);
     bold(text);
     italic(text);
     strike(text);
@@ -36,6 +46,31 @@ public class MarkDown {
       inputText.setSpan(new AbsoluteSizeSpan(0),
           matcher.start(), matcher.end(),
           Spanned.SPAN_EXCLUSIVE_EXCLUSIVE);
+    }
+  }
+
+  private static void highlightLink1(SpannableString inputText) {
+    Pattern linkPattern = Pattern.compile(
+        "\\[([^\\]]+)\\]\\(((?:http|https):\\/\\/[^\\)]+)\\)", Pattern.MULTILINE);
+    final Matcher matcher = linkPattern.matcher(inputText);
+    while (matcher.find()) {
+      ClickableSpan span = createLinkSpan(matcher.group(2));
+      setSpan(span, inputText,
+          matcher.start(), matcher.end(),
+          1, matcher.group(2).length() + 3);
+    }
+  }
+
+  private static void highlightLink2(SpannableString inputText) {
+    Pattern linkPattern = Pattern.compile(
+        "(?:<|&lt;)((?:http|https):\\/\\/[^\\|]+)\\|(.+?)(?=>|&gt;)(?:>|&gt;)", Pattern.MULTILINE);
+    Matcher matcher = linkPattern.matcher(inputText);
+    while (matcher.find()) {
+      ClickableSpan span = createLinkSpan(matcher.group(2));
+      setSpan(span, inputText,
+          matcher.start(), matcher.end(),
+          matcher.group(1).length() + matcher.group(2).length() + 1,
+          matcher.group(5).length() + matcher.group(6).length());
     }
   }
 
@@ -95,5 +130,29 @@ public class MarkDown {
           end - markEndLen, end,
           Spanned.SPAN_EXCLUSIVE_EXCLUSIVE);
     }
+  }
+
+  /*package*/ static ClickableSpan createLinkSpan(final String url) {
+    return new ClickableSpan() {
+      @Override
+      public void onClick(View view) {
+        final Context context = view.getContext();
+        try {
+          Intent intent = new Intent(Intent.ACTION_VIEW, Uri.parse(url));
+          intent.setFlags(Intent.FLAG_ACTIVITY_NEW_TASK);
+          context.startActivity(intent);
+          return;
+        } catch (Exception exception) {
+        }
+
+        try {
+          ClipboardManager clipboardManager =
+              (ClipboardManager) context.getSystemService(Context.CLIPBOARD_SERVICE);
+          clipboardManager.setPrimaryClip(ClipData.newPlainText("linkURL", url));
+          Toast.makeText(context, "Copied to clipboard", Toast.LENGTH_SHORT).show();
+        } catch (Exception exception) {
+        }
+      }
+    };
   }
 }
